% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalglmnet.R
\name{causalglmnet}
\alias{causalglmnet}
\title{causalglmnet
High dimensional semiparametric generalized linear models for causal inference using the LASSO.
Supports flexible semiparametric conditional average treatment effect (CATE), conditional odds ratio (OR), and conditional relative risk (RR) estimation
\code{\link[glmnet]{cv.glmnet}} is used to fit all nuisance parameters. The parametric component of the semiparametric model is not penalized.
This function is almost just a wrapper for \code{causalglm}.}
\usage{
causalglmnet(
  formula,
  data,
  W,
  A,
  Y,
  estimand = c("CATE", "OR", "RR"),
  max_degree = 1,
  cross_fit = TRUE,
  constant_variance_CATE = FALSE,
  weights = NULL,
  parallel = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{formula}{A R formula object specifying the parametric form of CATE, OR, or RR (depending on method).}

\item{data}{A data.frame or matrix containing the numeric values corresponding with the nodes \code{W}, \code{A} and \code{Y}.}

\item{W}{A character vector of covariates contained in \code{data}}

\item{A}{A character name for the treatment assignment variable contained in \code{data}}

\item{Y}{A character name for the outcome variable contained in \code{data} (outcome can be continuous, nonnegative or binary depending on method)}

\item{estimand}{Estimand/parameter to estimate. Choices are:
CATE: Estimate conditional average treatment effect with \code{\link[tmle3]{Param_spCATE}} assuming it satisfies parametric model \code{formula}.
OR: Estimate conditional odds ratio with \code{\link[tmle3]{Param_spOR}} assuming it satisfies parametric model \code{formula}.
RR: Estimate conditional relative risk with \code{\link[tmle3]{Param_spRR}} assuming it satisfies parametric model \code{formula}.}

\item{cross_fit}{Whether to cross-fit the initial estimator. This is always set to FALSE if argument \code{sl3_Learner} is provided.
learning_method = `SuperLearner` is always cross-fitted (default).
 learning_method = `xgboost` and `ranger` are always cross-fitted regardless of the value of \code{cross_fit}
 All other learning_methods are only cross-fitted if `cross_fit=TRUE`.
 Note, it is not necessary to cross-fit glm, glmnet, gam or mars as long as the dimension of W is not too high.
 In smaller samples and lower dimensions, it may fact hurt to cross-fit.}

\item{weights}{An optional vector of weights to use in procedure.}

\item{parallel}{See \code{\link[glmnet]{cv.glmnet}}}

\item{...}{Other arguments to pass to \code{\link[glmnet]{cv.glmnet}}}
}
\description{
causalglmnet
High dimensional semiparametric generalized linear models for causal inference using the LASSO.
Supports flexible semiparametric conditional average treatment effect (CATE), conditional odds ratio (OR), and conditional relative risk (RR) estimation
\code{\link[glmnet]{cv.glmnet}} is used to fit all nuisance parameters. The parametric component of the semiparametric model is not penalized.
This function is almost just a wrapper for \code{causalglm}.
}
