% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_msm}
\alias{plot_msm}
\title{Plotting of univariate marginal structural models}
\usage{
plot_msm(object, data = object$args$data)
}
\arguments{
\item{object}{A `msmglm` object}

\item{data}{The data containing `V` for plotting}
}
\description{
Plotting of univariate marginal structural models
}
