% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.RR}
\alias{sim.RR}
\title{Simulate a dataset with known constant RR}
\usage{
sim.RR(
  n = 1500,
  p = 2,
  prop_active = 1,
  formula_estimand = ~1,
  formula_A = ~.,
  formula_Y0W = ~.,
  beta = NULL,
  beta_A = NULL,
  beta_Y = NULL
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Dimension of W}

\item{prop_active}{Proportion of active (nonzero coef) variables}
}
\description{
Simulate a dataset with known constant RR
}
